moved spheroidal routines to spheroidal_gridding
