+++ src/mpol/utils.py
@@ -1,4 +1,5 @@
 import torch
+import numpy as np
 
 from .constants import arcsec, cc, kB
 
@@ -25,6 +26,27 @@ def get_Jy_arcsec2(T_b, nu=230e9):
     Jy_arcsec2 = Jy_ster * arcsec ** 2
 
     return Jy_arcsec2
+
+
+def fftspace(width, N):
+    """Delivers a (nearly) symmetric coordinate array that spans :math:`N` elements (where :math:`N` is even) from `-width` to `+width`, but ensures that the middle point lands on :math:`0`. The array indices go from :math:`0` to :math:`N -1.`
+    
+    Args:
+        width (float): the width of the array
+        N (int): the number of elements in the array
+        
+    Returns:
+        numpy.float64 1D array: the fftspace array
+    
+    """
+    assert N % 2 == 0, "N must be even."
+
+    dx = width * 2.0 / N
+    xx = np.empty(N, np.float)
+    for i in range(N):
+        xx[i] = -width + i * dx
+
+    return xx
 
 
 def fftshift(x, axes=None):
